truncatedb pstmt

+++ backend/skynet_backend/src/main/java/toctep/skynet/backend/dal/dao/impl/mysql/TweetKeywordDaoImpl.java
@@ -21,7 +21,7 @@ public class TweetKeywordDaoImpl extends TweetKeywordDao {
 		Param[] params = new Param[] {
 			new Param(tweetKeyword.getTweet().getId(), Types.BIGINT),
 			new Param(tweetKeyword.getTweetKeywordValue(), Types.VARCHAR),
-			new Param(tweetKeyword.getKeyword().getId(), Types.BIGINT)
+			new Param(tweetKeyword.getKeyword().getId(), Types.INTEGER)
 		};
 			
 		int id = MySqlUtil.getInstance().insert(query, params);
@@ -68,6 +68,7 @@ public class TweetKeywordDaoImpl extends TweetKeywordDao {
 			tweetKeyword.setKeyword(Keyword.select((Integer) record.get(2)));
 			tweetKeywords.add(tweetKeyword);
 		}
+		
 		return tweetKeywords;
 	}
